Organize input-related files into src/lib/game/input/ subdirectory #43
