+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@joshuafolkken/game-kit",
-	"version": "0.13.0",
+	"version": "0.14.0",
 	"keywords": [
 		"svelte"
 	],
