+++ src/lib/game/input/pointer-compute.svelte.spec.ts
