+++ src/lib/game/input/pointer-compute.ts
